code clean up and organized
